fix: refatorando users e kids com roles e permissions em andamento

+++ app/Http/Controllers/KidsController.php
@@ -13,12 +13,13 @@ use Exception;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use Yajra\DataTables\DataTables;
+use Spatie\Permission\Models\Role as SpatieRole;
 
 class KidsController extends Controller
 {
     public function index()
     {
-        $message = label_case('Index Kids ').' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+        $message = label_case('Index Kids ') . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
         Log::debug($message);
 
         return view('kids.index');
@@ -27,24 +28,19 @@ class KidsController extends Controller
     public function index_data()
     {
 
-        $data = Kid::getKids();
-
-
-        foreach ($data as $kid) {
-            $this->authorize('view', $kid);
-        }
+        $data = Kid::All();
 
         return Datatables::of($data)
             ->addColumn('action', function ($data) {
 
-                if (request()->user()->can('kids.update') || request()->user()->can('kids.store')) {
+                if (request()->user()->can('edit kids')) {
 
                     $html = '<div class="dropdown">';
                     $html .= '<button class="btn btn-sm btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false"><i class="bi bi-gear"></i></button><ul class="dropdown-menu" aria-labelledby="dropdownMenuButton1">';
-                    $html .= '<li><a class="dropdown-item" href="'.route('kids.edit', $data->id).'"><i class="bi bi-pencil"></i> Editar</a></li>';
+                    $html .= '<li><a class="dropdown-item" href="' . route('kids.edit', $data->id) . '"><i class="bi bi-pencil"></i> Editar</a></li>';
 
                     if ($data->checklists()->count() > 0) {
-                        $html .= '<li><a class="dropdown-item" href="'.route('kids.show', $data->id).'"><i class="bi bi-check2-square"></i> Checklist</a></li>';
+                        $html .= '<li><a class="dropdown-item" href="' . route('kids.show', $data->id) . '"><i class="bi bi-check2-square"></i> Checklist</a></li>';
                     }
 
                     $html .= '</ul></div>';
@@ -59,7 +55,7 @@ class KidsController extends Controller
                 return $data->birth_date;
             })
             ->editColumn('checklists', function ($data) {
-                return '<span class="badge bg-success"><i class="bi bi-check"></i> '.$data->checklists->count().' Checklist(s) </span>';
+                return '<span class="badge bg-success"><i class="bi bi-check"></i> ' . $data->checklists->count() . ' Checklist(s) </span>';
             })
             ->editColumn('profession_id', function ($data) {
                 // Exibe o nome do profissional ou 'Não atribuído' caso não tenha um profissional
@@ -76,13 +72,12 @@ class KidsController extends Controller
 
     public function create()
     {
-        $message = label_case('Create Kids').' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+        $this->authorize('create', Kid::class);
+
+        $message = label_case('Create Kids') . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
         Log::info($message);
 
-        $responsibles = User::where('role_id', User::ROLE_PAIS)->get();
-        $professions = User::where('role_id', User::ROLE_PROFESSION)->get();
-
-        return view('kids.create', compact('professions', 'responsibles'));
+        return view('kids.create');
     }
 
     public function store(KidRequest $request)
@@ -91,52 +86,29 @@ class KidsController extends Controller
 
         DB::beginTransaction();
         try {
-            $message = label_case('Store Kids '.self::MSG_CREATE_SUCCESS).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            $message = label_case('Store Kids ' . self::MSG_CREATE_SUCCESS) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::info($message);
 
-            /*
-            // cadastra user com role_id = 3 (pais)
-            $userData = [
-                'name' => $request->responsible_name,
-                'email' => $request->email,
-                'phone' => $request->phone,
-                'postal_code' => $request->cep,
-                'street' => $request->logradouro,
-                'number' => $request->numero,
-                'complement' => $request->complemento,
-                'neighborhood' => $request->bairro,
-                'city' => $request->cidade,
-                'state' => $request->estado,
-                'password' => bcrypt('default_password'), // Ou você pode gerar uma senha aleatória
-                'role_id' => 3, // ROLE_PAIS (assumindo que 3 corresponde a ROLE_PAIS)
-                'created_by' => auth()->user()->id,
-            ];
-            $user = User::create($userData);
-            Log::info('User created: '.$user->id. ' created by: '.auth()->user()->id);
-            */
 
             $kidData = [
                 'name' => $request->name,
                 'birth_date' => $request->birth_date,
             ];
 
-            if(auth()->user()->isProfessional()){
-                $kidData['profession_id'] = auth()->user()->id;
-            }
+            // removido is professional
 
             $kid = Kid::create($kidData);
-            Log::info('Kid created: '.$kid->id.' created by: '.auth()->user()->id);
+            Log::info('Kid created: ' . $kid->id . ' created by: ' . auth()->user()->id);
 
             flash(self::MSG_CREATE_SUCCESS)->success();
 
             DB::commit();
 
             return redirect()->route('kids.index');
-
         } catch (Exception $e) {
             DB::rollBack();
             flash(self::MSG_CREATE_ERROR)->warning();
-            $message = label_case('Store Kids '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            $message = label_case('Store Kids ' . $e->getMessage()) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::error($message);
 
             return redirect()->route('kids.index');
@@ -174,7 +146,7 @@ class KidsController extends Controller
             return view('kids.show', $data);
         } catch (Exception $e) {
             flash(self::MSG_NOT_FOUND)->warning();
-            $message = label_case('Show Kids '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            $message = label_case('Show Kids ' . $e->getMessage()) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::error($message);
 
             return redirect()->back();
@@ -183,17 +155,39 @@ class KidsController extends Controller
 
     public function edit(Kid $kid)
     {
+        $this->authorize('update', $kid);
+
         try {
-            $message = label_case('Edit Kids ').' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            $message = label_case('Edit Kids ') . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::info($message);
 
-            $responsibles = User::where('role_id', User::ROLE_PAIS)->get();
-            $professions = User::where('role_id', User::ROLE_PROFESSION)->get();
+            // Verificando se o papel 'pais' existe, caso contrário lançar exceção
+            $parentRole = SpatieRole::where('name', 'pais')->first();
+            if (!$parentRole) {
+                throw new Exception("O papel 'pais' não existe no sistema.");
+            }
+
+            // Verificando se o papel 'professional' existe, caso contrário lançar exceção
+            $professionalRole = SpatieRole::where('name', 'professional')->first();
+            if (!$professionalRole) {
+                throw new Exception("O papel 'professional' não existe no sistema.");
+            }
+
+            // Buscando usuários com o papel 'pais'
+            $responsibles = User::whereHas('roles', function ($query) {
+                $query->where('name', 'pais');
+            })->get();
+
+            // Buscando usuários com o papel 'professional'
+            $professions = User::whereHas('roles', function ($query) {
+                $query->where('name', 'professional');
+            })->get();
+
 
             return view('kids.edit', compact('kid', 'responsibles', 'professions'));
         } catch (Exception $e) {
-            flash(self::MSG_NOT_FOUND)->warning();
-            $message = label_case('Update Kids '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            flash($e->getMessage())->warning();
+            $message = label_case('Update Kids ' . $e->getMessage()) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::error($message);
 
             return redirect()->back();
@@ -202,46 +196,63 @@ class KidsController extends Controller
 
     public function update(KidRequest $request, Kid $kid)
     {
+        // Verifica se o usuário está autorizado a atualizar o registro
         $this->authorize('update', $kid);
 
         try {
-            $message = label_case('Update Kids '.self::MSG_UPDATE_SUCCESS).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            // Loga a tentativa de atualização
+            $message = label_case('Update Kids ' . self::MSG_UPDATE_SUCCESS) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::info($message);
 
+            // Coleta os dados do request
             $data = $request->all();
-            if(auth()->user()->isProfessional()){
-                $data['profession_id'] = auth()->user()->id;
-            }
 
+            // Removida a verificação de papel 'isProfessional'
+            // O tratamento para usuários profissionais será feito futuramente
+
+            // Atualiza os dados da criança
             $kid->update($data);
-            //SendKidUpdateJob::dispatch($kid)->onQueue('emails');
 
+            // Opcional: Disparar job de atualização de criança
+            // SendKidUpdateJob::dispatch($kid)->onQueue('emails');
+
+            // Mensagem de sucesso
             flash(self::MSG_UPDATE_SUCCESS)->success();
 
+            // Redireciona para a página de edição
             return redirect()->route('kids.edit', $kid->id);
         } catch (Exception $e) {
-            $message = label_case('Update Kids Error'.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            // Loga o erro em caso de falha
+            $message = label_case('Update Kids Error' . $e->getMessage()) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::error($message);
 
+            // Mensagem de erro
             flash(self::MSG_UPDATE_ERROR)->warning();
 
+            // Redireciona de volta para a página anterior
             return redirect()->back();
         }
     }
 
+
     public function destroy(Kid $kid)
     {
+        $this->authorize('delete', $kid);
         try {
-            $message = label_case('Destroy Kids '.self::MSG_DELETE_SUCCESS).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            $message = label_case('Destroy Kids ' . self::MSG_DELETE_SUCCESS) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::info($message);
+
+
             $kid->delete();
             flash(self::MSG_DELETE_SUCCESS)->success();
 
             return redirect()->route('kids.index');
         } catch (Exception $e) {
-            flash(self::MSG_NOT_FOUND)->warning();
-            $message = label_case('Destroy Kids '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+
+            $message = label_case('Destroy Kids ' . $e->getMessage()) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::error($message);
+
+            flash(self::MSG_NOT_FOUND)->warning();
 
             return redirect()->back();
         }
@@ -292,13 +303,13 @@ class KidsController extends Controller
 
                     $pdf->Ln(5);
                     $pdf->SetFont('helvetica', 'B', 10);
-                    $txt = $competence->level_id.$v['domain']->initial.$competence->code.' - '.$competence->description;
+                    $txt = $competence->level_id . $v['domain']->initial . $competence->code . ' - ' . $competence->description;
                     $pdf->Ln(5);
                     $pdf->Write(0, $txt, '', 0, 'L', true);
 
                     $pdf->Ln(1);
                     $pdf->SetFont('helvetica', 'I', 8);
-                    $pdf->Write(0, '"'.$competence->description_detail.'"', '', 0, 'L', true);
+                    $pdf->Write(0, '"' . $competence->description_detail . '"', '', 0, 'L', true);
 
                     $pdf->Ln(4);
                     $pdf->SetFont('helvetica', '', 9);
@@ -311,10 +322,9 @@ class KidsController extends Controller
                 //                }
             }
 
-            $pdf->Output($nameKid.'_'.$date->format('dmY').'_'.$plane->id.'.pdf', 'I');
-
+            $pdf->Output($nameKid . '_' . $date->format('dmY') . '_' . $plane->id . '.pdf', 'I');
         } catch (Exception $e) {
-            $message = label_case('Plane Kids '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            $message = label_case('Plane Kids ' . $e->getMessage()) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::error('Exibe Plano Erro', [
                 'message' => $e->getMessage(),
                 'file' => $e->getFile(),
@@ -352,9 +362,9 @@ class KidsController extends Controller
         $pdf->setViewerPreferences($preferences);
         $pdf->AddPage();
         $pdf->SetFont('helvetica', '', 18);
-        $pdf->Cell(0, 75, 'PLANO DE INTERVENÇÃO N.: '.$plane_id, 0, 1, 'C');
+        $pdf->Cell(0, 75, 'PLANO DE INTERVENÇÃO N.: ' . $plane_id, 0, 1, 'C');
         $pdf->SetFont('helvetica', '', 12);
-        $txt = 'Terapeuta: '.$therapist.',  Data: '.$date;
+        $txt = 'Terapeuta: ' . $therapist . ',  Data: ' . $date;
         $pdf->Cell(0, 2, $txt, 0, 1, 'C');
         $pdf->Ln(15);
         $pdf->SetFont('helvetica', '', 18);
